+++ src/main/java/org/jabref/logic/exporter/BibDatabaseWriter.java
@@ -322,7 +322,7 @@ public abstract class BibDatabaseWriter {
         List<FieldChange> changes = new ArrayList<>();
         CitationKeyGenerator keyGenerator = new CitationKeyGenerator(databaseContext, preferences.getCitationKeyPatternPreferences());
         for (BibEntry bes : entries) {
-            Optional<String> oldKey = bes.getCiteKeyOptional();
+            Optional<String> oldKey = bes.getCitationKey();
             if (StringUtil.isBlank(oldKey)) {
                 Optional<FieldChange> change = keyGenerator.generateAndSetKey(bes);
                 change.ifPresent(changes::add);