+++ src/test/java/org/jabref/logic/importer/fetcher/GoogleScholarTest.java
@@ -61,7 +61,7 @@ class GoogleScholarTest implements SearchBasedFetcherCapabilityTest {
     @DisabledOnCIServer("CI server is blocked by Google")
     void findSingleEntry() throws FetcherException {
         entry.setType(StandardEntryType.InProceedings);
-        entry.setCiteKey("geiger2013detecting");
+        entry.setCitationKey("geiger2013detecting");
         entry.setField(StandardField.TITLE, "Detecting Interoperability and Correctness Issues in BPMN 2.0 Process Models.");
         entry.setField(StandardField.AUTHOR, "Geiger, Matthias and Wirtz, Guido");
         entry.setField(StandardField.BOOKTITLE, "ZEUS");